pyproject.toml
[project]
dependencies = [
  "anyio>=4.4.0",
  "beautifulsoup4>=4.12.3",
  "certifi>=2024.7.4",
  "charset-normalizer>=3.3.2",
  "h11>=0.14.0",
  "httpcore>=1.0.5",
  "httpx>=0.27.0",
  "idna>=3.8",
  "python-dotenv>=1.0.1",
  "python-telegram-bot>=21.4",
  "sniffio>=1.3.1",
  "soupsieve>=2.6",
  "urllib3>=2.2.2",
  "requests>=2.32.3",
  "mysql-connector-python>=9.0.0",
  "structlog>=24.4.0",
  "google-auth",
  "google-auth-oauthlib",
  "google-auth-httplib2",
  "google-api-python-client",
  "faker>=33.0.0",
  "kink>=0.8.1",
  "pyyaml>=6.0.2",
  "pytest>=8.3.3",
  "pytest-asyncio>=0.24.0",
  "telethon>=1.38.1",
]
description = "A Telegram bot for the SysArmy Galicia group."
name = "telegram-bot"
readme = "README.md"
requires-python = ">=3.10"
version = "0.1.0"

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[tool.uv]
dev-dependencies = [
  "ruff>=0.6.2",
]
